style(utils) rename abstract
